feat: core-explorer for core-v3
